requirements.txt
streamlit==0.69.1
numpy==1.19.1
plotly==4.11.0
pytorch torchvision cpuonly -c pytorch
fastai==1.0.61